requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic>=2.4.1,<2.6 